+++ app/Components/Sudoku Board/Components/Puzzle.tsx
@@ -1,9 +1,6 @@
 import { View } from "react-native";
 import React from "react";
-import {
-  CellProps,
-  SudokuObjectProps,
-} from "../../../Functions/LocalStore/DataStore/LocalDatabase";
+import { CellProps, SudokuObjectProps } from "../../../Functions/LocalDatabase";
 
 interface PuzzleProps {
   renderCell: (
