+++ app/Components/Sudoku Board/Functions/generateGame.ts
@@ -1,5 +1,5 @@
-import { Puzzles } from "../../../Functions/Api/Puzzles";
-import { SudokuObjectProps } from "../../../Functions/LocalStore/DataStore/LocalDatabase";
+import { Puzzles } from "../../../Api/Puzzles";
+import { SudokuObjectProps } from "../../../Functions/LocalDatabase";
 import { SudokuBoardProps } from "../SudokuBoard";
 
 export async function generateGame(props: SudokuBoardProps) {
