+++ app/Functions/learnedLessons.ts
