+++ app/Pages/ProfilePage.tsx
@@ -3,7 +3,7 @@ import { useWindowDimensions, View, ScrollView } from "react-native";
 import { SafeAreaProvider, SafeAreaView } from "react-native-safe-area-context";
 import { Switch, Text, useTheme } from "react-native-paper";
 import { PreferencesContext } from "../Contexts/PreferencesContext";
-import { formatLessonNameArray } from "../Functions/ContextParsing/learnedLessons";
+import { formatLessonNameArray } from "../Functions/learnedLessons";
 
 const ProfilePage = () => {
   const theme = useTheme();
